Refactor RequestStartedCallback out of GatewayBridge
